bags corrected + statistic shows only for keys that was typed >= 3times
